test: ensure core.load_solver import
