+++ README.md
@@ -138,7 +138,7 @@ NumPy-style docstrings.
 |-------------|-------------|-------------|
 | `bsde_dsgE.core.Solver` | Base BSDE solver | [`bsde_dsgE/core/solver.py`](bsde_dsgE/core/solver.py) |
 | `bsde_dsgE.core.ResNet` | Minimal residual network | [`bsde_dsgE/core/nets.py`](bsde_dsgE/core/nets.py) |
-| `bsde_dsgE.core.load_solver` | Factory for ``Solver`` and ``ResNet`` | [`bsde_dsgE/core/init.py`](bsde_dsgE/core/init.py) |
+| `bsde_dsgE.core.load_solver` | Factory for ``Solver`` and ``ResNet`` | [`bsde_dsgE/core/__init__.py`](bsde_dsgE/core/__init__.py) |
 | `bsde_dsgE.kfac.KFACPINNSolver` | KFAC training loop | [`bsde_dsgE/kfac/solver.py`](bsde_dsgE/kfac/solver.py) |
 | `bsde_dsgE.kfac.kfac_update` | Single KFAC step | [`bsde_dsgE/kfac/optimizer.py`](bsde_dsgE/kfac/optimizer.py) |
 | `bsde_dsgE.kfac.poisson_1d_residual` | 1‑D Poisson residual | [`bsde_dsgE/kfac/pde.py`](bsde_dsgE/kfac/pde.py) |
